+++ src/components/basic_table/basic_table.tsx
@@ -225,6 +225,7 @@ export type EuiBasicTableProps<T> = CommonProps &
   (BasicTableProps<T> | BasicTableWithPaginationProps<T>);
 
 interface State<T> {
+  initialSelectionRendered: boolean;
   selection: T[];
 }
 
@@ -284,12 +285,18 @@ export class EuiBasicTable<T = any> extends Component<
   private cleanups: Array<() => void> = [];
   private tbody: HTMLTableSectionElement | null = null;
 
-  state = {
-    selection: [],
-  };
+  constructor(props: EuiBasicTableProps<T>) {
+    super(props);
+    this.state = {
+      // used for checking if  initial selection is rendered
+      initialSelectionRendered: false,
+      selection: [],
+    };
+  }
 
   componentDidMount() {
     if (this.props.loading && this.tbody) this.addLoadingListeners(this.tbody);
+    this.getInitialSelection();
   }
 
   componentDidUpdate(prevProps: EuiBasicTableProps<T>) {
@@ -300,10 +307,27 @@ export class EuiBasicTable<T = any> extends Component<
         this.removeLoadingListeners();
       }
     }
+    this.getInitialSelection();
   }
 
   componentWillUnmount() {
     this.removeLoadingListeners();
+  }
+
+  getInitialSelection() {
+    if (
+      this.props.selection &&
+      this.props.selection.initialSelected &&
+      !this.state.initialSelectionRendered &&
+      this.props.items.length > 0
+    ) {
+      this.setState({ selection: this.props.selection.initialSelected });
+      this.setState({ initialSelectionRendered: true });
+    }
+  }
+
+  setSelection(newSelection: T[]) {
+    this.changeSelection(newSelection);
   }
 
   private setTbody = (tbody: HTMLTableSectionElement | null) => {